hw/intc/aspeed: Add ID to trace events for better debugging
Currently, it is difficult to recognize whether these trace events are from
INTC0 or INTC1. To make these trace events more readable, add an ID to the
INTC trace events.
Updated trace events to include the "id" field for better identification.
Updated the "AspeedINTCClass" structure to include an "id" field.

Signed-off-by: Jamin Lin <jamin_lin@aspeedtech.com>
Change-Id: Iaffb69515ed7ca2b448c572e4c80be27e10ecddb